Enable cache & Update fip tool
1. Fix uboot cache code.
2. Update bl2 cache code.
3. Update fip_create tool, align each bin to 16KB.

+++ arch/arm/cpu/armv8/cache_v8.c
@@ -25,7 +25,7 @@ void set_pgtable_section(u64 *page_table, u64 index, u64 section,
 /* to activate the MMU we need to set up virtual memory */
 static void mmu_setup(void)
 {
-	int i, j, el;
+	u64 i, j, el;
 	bd_t *bd = gd->bd;
 	u64 *page_table = (u64 *)gd->arch.tlb_addr;
